Commands refactor + other fixes and minor shit (#68)
* refactor(commands): GIANT REFACTOR

split Commands into components and then abuse multiple inheritance to stich it all back together into the same as before

also switched to poetry for build stuff

* revert: uh oops that's me-specific

* Update pyproject.toml

* fix: missing awaits causing failure to respond

* feat(help): usage footer

* feat(help): delete help request message if in public channel

* build: make poetry install cachable

separate out the install dependencies and install self - this works because docker sees that the toml didn't change and uses the cached install
when .fred is copied, it is properly installed with your changes

* refactor(init): error channel in env vars

* chore: version bump

* fix: ensure .env is present and make default value for ERROR_CHANNEL

* fix: indentation issues displayed on mobile

* fix: don't die if .env isn't present

i can't believe i am committing a single character
