+++ fred/cogs/welcome.py
@@ -1,7 +1,7 @@
-import nextcord.ext.commands as commands
+from nextcord.ext import commands
 
-import config
-from libraries import common
+from .. import config
+from ..libraries import common
 
 
 class Welcome(commands.Cog):
