+++ pyproject.toml
@@ -1,0 +1,24 @@
+[tool.poetry]
+name = "FICSIT-Fred"
+version = "2.18.0"
+description = "A Discord bot for the Satisfactory Modding Discord "
+authors = ["Feyko <feykook@gmail.com>", "Mircea Roata <mircearoatapalade@yahoo.ro>", "Astro Floof <does.not.kompute@gmail.com>"]
+license = "MIT License"
+
+[tool.poetry.dependencies]
+python = "^3.10"
+nextcord = "^2.0.0-alpha.6"
+Pillow = "^9.0.0"
+pytesseract = "^0.3.8"
+psycopg2-binary = "^2.9.3"
+python-dotenv = "^0.19.2"
+jsonpickle = "^2.1.0"
+algoliasearch = "^2.6.1"
+SQLObject = "^3.9.1"
+google-cloud-dialogflow = "^2.11.0"
+
+[tool.poetry.dev-dependencies]
+
+[build-system]
+requires = ["poetry-core>=1.0.0"]
+build-backend = "poetry.core.masonry.api"
